added warning for when you sign a step with remaining contributions
