+++ src/apps/ScopeChangeRequest/Components/SearchableDropdown/PCSPersonSearch.tsx
@@ -1,8 +1,5 @@
-import { Icon } from '@equinor/eds-core-react';
-import { tokens } from '@equinor/eds-tokens';
 import { SingleValue, Theme } from 'react-select';
 import AsyncSelect from 'react-select/async';
-import { useHttpClient } from '../../../../Core/Client/Hooks/useApiClient';
 import { searchPcs } from '../../Api/Search/PCS/searchPcs';
 import { applyEdsComponents, applyEdsStyles, applyEDSTheme } from './applyEds';
 
@@ -17,10 +14,8 @@ interface PCSLinkProps {
 }
 
 export const PCSPersonSearch = ({ person, setPerson }: PCSLinkProps): JSX.Element => {
-    const { procosys } = useHttpClient();
-
     const loadOptions = async (inputValue: string, callback: (options: SelectOption[]) => void) => {
-        callback(await searchPcs(inputValue, 'person', procosys));
+        callback(await searchPcs(inputValue, 'person'));
     };
 
     return (
@@ -32,7 +27,7 @@ export const PCSPersonSearch = ({ person, setPerson }: PCSLinkProps): JSX.Elemen
                     defaultOptions={false}
                     value={person}
                     styles={applyEdsStyles()}
-                    controlShouldRenderValue={false}
+                    controlShouldRenderValue={true}
                     components={{ ...applyEdsComponents() }}
                     placeholder={`Type to search..`}
                     noOptionsMessage={(obj: { inputValue: string }) => {
